refactor(DSL): split out various dsls
