+++ packages/core/src/Prelude/DSL/index.ts
@@ -1,9 +1,15 @@
 // ets_tracing: off
 
-export * from "./dsl"
-export * from "./validation"
-export * from "./applicative"
-export * from "./gen"
+// codegen:start {preset: barrel, include: ./*.ts}
+export * from "./access-provide"
 export * from "./alternative"
-export * from "./matchers"
+export * from "./applicative"
+export * from "./apply"
+export * from "./chain"
 export * from "./conditionals"
+export * from "./do"
+export * from "./gen"
+export * from "./matchers"
+export * from "./succeed"
+export * from "./validation"
+// codegen:end
